requirements.txt
streamlit==1.37.1
pandas==2.2.2
psycopg2-binary==2.9.9
plotly==5.24.0
streamlit-plotly-events==0.0.6
xlsxwriter==3.2.0
reportlab==4.2.2